+++ packages/@runlightyear/lightyear/src/index.ts
@@ -7,28 +7,33 @@
 /** Base Functions **/
 
 /* Action */
-import { defineAction } from "./base/action";
-import type { DefineActionProps } from "./base/action";
+export { defineAction } from "./base/action";
+export type { DefineActionProps } from "./base/action";
 
 /* Auth */
-import { getAuthData, updateAuthData, updateAuthDataState } from "./base/auth";
-import type { AuthData, AuthProps } from "./base/auth";
+export {
+  getAuthData,
+  updateAuthData,
+  updateAuthDataState,
+  setAuthError,
+} from "./base/auth";
+export type { AuthData, AuthProps } from "./base/auth";
 
 /* Base Request */
-import baseRequest from "./base/baseRequest";
-import type { BaseRequestProps } from "./base/baseRequest";
+export { default as baseRequest } from "./base/baseRequest";
+export type { BaseRequestProps } from "./base/baseRequest";
 
 /* Cache */
-import { getCache, saveCache } from "./base/cache";
-import type { GetCacheProps, SaveCacheProps, CacheData } from "./base/cache";
+export { getCache, saveCache } from "./base/cache";
+export type { GetCacheProps, SaveCacheProps, CacheData } from "./base/cache";
 
 /* Deploy */
-import { deploy, getDeployData } from "./base/deploy";
-import type { DeployFuncProps, DeployFunc } from "./base/deploy";
+export { deploy, getDeployData } from "./base/deploy";
+export type { DeployFuncProps, DeployFunc } from "./base/deploy";
 export { getDeployList } from "./base/deploy";
 
 /* Http Proxy */
-import type {
+export type {
   HttpProxyRequestProps,
   HttpProxyResponse,
   HttpProxyResponseError,
@@ -39,28 +44,29 @@ export { defineOAuth } from "./base/oauth";
 export type { DefineOAuthProps } from "./base/oauth";
 
 /* Runs */
-import { run } from "./run";
-import type { RunFunc, RunFuncProps } from "./run";
+export { run } from "./run";
+export type { RunFunc, RunFuncProps } from "./run";
 // TODO: where should these live?
-import { PollingData, WebhookDeliveryData } from "./base/runData";
+export type { PollingData, WebhookDeliveryData } from "./base/runData";
 
 /* Secrets */
-import { setSecret } from "./base/secret";
-import type { SecretData } from "./base/secret";
+export { setSecret } from "./base/secret";
+export type { SecretData } from "./base/secret";
 
 /* Smtp Proxy */
-import type {
+export type {
   SmtpProxyRequestProps,
   SmtpProxyResponse,
   EmailAddress,
 } from "./base/smtp";
 
 /* Variables */
-import { getVariableData, setVariable, VariableData } from "./base/variable";
+export { getVariableData, setVariable } from "./base/variable";
+export type { VariableData } from "./base/variable";
 
 /* Webhooks */
-import { defineWebhook, getWebhookData } from "./base/webhook";
-import type {
+export { defineWebhook, getWebhookData } from "./base/webhook";
+export type {
   DefineWebhookProps,
   SubscribePropsFunc,
   SubscribePropsFuncProps,
@@ -73,51 +79,51 @@ import type {
 /** Connectors **/
 
 /* Base Connector */
-import { BaseConnector } from "./connectors/BaseConnector";
-import type { BaseConnectorProps } from "./connectors/BaseConnector";
+export { BaseConnector } from "./connectors/BaseConnector";
+export type { BaseConnectorProps } from "./connectors/BaseConnector";
 
 /* Auth Connector */
-import { AuthConnector } from "./connectors/AuthConnector";
-import type { AuthConnectorProps } from "./connectors/AuthConnector";
+export { AuthConnector } from "./connectors/AuthConnector";
+export type { AuthConnectorProps } from "./connectors/AuthConnector";
 
 /* OAuth Connector */
-import { OAuthConnector } from "./connectors/OAuthConnector";
-import type {
+export { OAuthConnector } from "./connectors/OAuthConnector";
+export type {
   OAuthConnectorProps,
   OAuthConfigData,
 } from "./connectors/OAuthConnector";
 
 /* Rest Connector */
-import { RestConnector } from "./connectors/RestConnector";
-import type { RestConnectorProps } from "./connectors/RestConnector";
+export { RestConnector } from "./connectors/RestConnector";
+export type { RestConnectorProps } from "./connectors/RestConnector";
 
 /* GraphQL Connector */
-import { GraphQLConnector } from "./connectors/GraphQLConnector";
-import type {
+export { GraphQLConnector } from "./connectors/GraphQLConnector";
+export type {
   GraphQLConnectorProps,
   GraphQLConnectorQueryProps,
 } from "./connectors/GraphQLConnector";
 
 /* Smtp Connector */
-import { SmtpConnector } from "./connectors/SmtpConnector";
-import { SmtpConnectorProps } from "./connectors/SmtpConnector";
+export { SmtpConnector } from "./connectors/SmtpConnector";
+export type { SmtpConnectorProps } from "./connectors/SmtpConnector";
 
 /** Logging **/
-import { PrefixedRedactedConsole } from "./logging/PrefixedRedactedConsole";
+export { PrefixedRedactedConsole } from "./logging/PrefixedRedactedConsole";
 
 /** Utilities **/
-import { base64ToBase64Url } from "./util/base64ToBase64Url";
-import camelize from "./util/camelize";
-import deCamelize from "./util/deCamelize";
-import { getApiKey } from "./util/getApiKey";
-import { getBaseUrl } from "./util/getBaseUrl";
-import { getEnvName } from "./util/getEnvName";
-import toTimestamp from "./util/toTimestamp";
+export { base64ToBase64Url } from "./util/base64ToBase64Url";
+export { default as camelize } from "./util/camelize";
+export { default as deCamelize } from "./util/deCamelize";
+export { getApiKey } from "./util/getApiKey";
+export { getBaseUrl } from "./util/getBaseUrl";
+export { getEnvName } from "./util/getEnvName";
+export { default as toTimestamp } from "./util/toTimestamp";
 export { dayjsUtc } from "./util/dayjsUtc";
 export { isFunction } from "./util/isFunction";
 
 /** Handler **/
-import { handler } from "./handler";
+export { handler } from "./handler";
 
 /** Action Results */
 
@@ -126,65 +132,4 @@ import { handler } from "./handler";
  *
  * This can be thrown inside a RunFunc to end execution and mark the run as Skipped.
  */
-const SKIPPED = "SKIPPED";
-
-export {
-  defineAction,
-  getAuthData,
-  getVariableData,
-  defineWebhook,
-  getWebhookData,
-  BaseConnector,
-  AuthConnector,
-  OAuthConnector,
-  RestConnector,
-  GraphQLConnector,
-  SmtpConnector,
-  deploy,
-  run,
-  PrefixedRedactedConsole,
-  handler,
-  toTimestamp,
-  camelize,
-  deCamelize,
-  HttpProxyResponseError,
-  SKIPPED,
-  getApiKey,
-  getBaseUrl,
-  getEnvName,
-  base64ToBase64Url,
-  setVariable,
-  setSecret,
-};
-
-export type {
-  AuthData,
-  AuthProps,
-  DefineActionProps,
-  VariableData,
-  SecretData,
-  BaseConnectorProps,
-  AuthConnectorProps,
-  HttpProxyRequestProps,
-  HttpProxyResponse,
-  SmtpProxyRequestProps,
-  SmtpProxyResponse,
-  EmailAddress,
-  OAuthConnectorProps,
-  OAuthConfigData,
-  RestConnectorProps,
-  GraphQLConnectorProps,
-  GraphQLConnectorQueryProps,
-  SmtpConnectorProps,
-  PollingData,
-  WebhookDeliveryData,
-  RunFunc,
-  RunFuncProps,
-  DefineWebhookProps,
-  SubscribePropsFunc,
-  SubscribePropsFuncProps,
-  SubscribeFunc,
-  SubscribeFuncProps,
-  UnsubscribeFunc,
-  UnsubscribeFuncProps,
-};
+export const SKIPPED = "SKIPPED";