+++ packages/@runlightyear/zoom/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/zoom
 
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.28.0
+
 ## 0.4.0
 
 ### Minor Changes